SlicePolygon : int형으로 cast 해서 보내주기

+++ SlicePolygon/SlicePolygon/polygon.h
@@ -22,7 +22,7 @@ public:
 
 	~Poly();
 
-	size_t getVerticesSize() { return vertices.size(); }
+	int getVerticesSize() { return static_cast<int>(vertices.size()); }
 	
 	glm::vec2 getVertex2(int index);
 	glm::vec4 getBoundBox();
